pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnirag"
version = "4.0.0a1"
description = "OmniRAG v4 — Open-source control plane for RAG systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "OmniRAG Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "click>=8.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "structlog>=24.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
langchain = ["langchain>=0.2", "langchain-community>=0.2"]
llamaindex = ["llama-index>=0.10"]
haystack = ["haystack-ai>=2.0"]
qdrant = ["qdrant-client>=1.7"]
huggingface = ["sentence-transformers>=2.5"]
all = [
    "omnirag[langchain,llamaindex,haystack,qdrant,huggingface]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
    "bandit>=1.7",
    "pre-commit>=3.6",
    "hypothesis>=6.98",
]

[project.scripts]
omnirag = "omnirag.cli.main:cli"

[project.urls]
Homepage = "https://github.com/RachEma-ux/omnirag"
Repository = "https://github.com/RachEma-ux/omnirag"
Issues = "https://github.com/RachEma-ux/omnirag/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["omnirag"]

[tool.coverage.report]
fail_under = 85
